Matrix: made dimension and number of nonzeros size_t (unsigned integer)

+++ uno/linear_algebra/COOSymmetricMatrix.cpp
@@ -7,28 +7,28 @@
  * https://en.wikipedia.org/wiki/Sparse_matrix#Coordinate_list_(COO)
  */
 
-COOSymmetricMatrix::COOSymmetricMatrix(int dimension, size_t capacity) : SymmetricMatrix(dimension, capacity), matrix(capacity),
+COOSymmetricMatrix::COOSymmetricMatrix(size_t dimension, size_t capacity) : SymmetricMatrix(dimension, capacity), matrix(capacity),
 row_indices(capacity), column_indices(capacity) {
 }
 
-COOSymmetricMatrix::COOSymmetricMatrix(int dimension, std::vector<double> matrix, std::vector<int> row_indices, std::vector<int> column_indices) :
+COOSymmetricMatrix::COOSymmetricMatrix(size_t dimension, std::vector<double> matrix, std::vector<int> row_indices, std::vector<int> column_indices) :
       SymmetricMatrix(dimension, matrix.capacity()), matrix(std::move(matrix)), row_indices(std::move(row_indices)),
       column_indices(std::move(column_indices)) {
 }
 
 // generic iterator
 void COOSymmetricMatrix::for_each(const std::function<void(int, int, double)>& f) const {
-   for (int k = 0; k < this->number_nonzeros; k++) {
+   for (size_t k = 0; k < this->number_nonzeros; k++) {
       int i = this->row_indices[k];
       int j = this->column_indices[k];
       f(i, j, this->matrix[k]);
    }
 }
 
-void COOSymmetricMatrix::insert(double term, int row_index, int column_index) {
+void COOSymmetricMatrix::insert(double term, size_t row_index, size_t column_index) {
    this->matrix[this->number_nonzeros] = term;
-   this->row_indices[this->number_nonzeros] = row_index;
-   this->column_indices[this->number_nonzeros] = column_index;
+   this->row_indices[this->number_nonzeros] = (int) row_index;
+   this->column_indices[this->number_nonzeros] = (int) column_index;
    this->number_nonzeros++;
 }
