+++ uno/linear_algebra/COOSymmetricMatrix.hpp
@@ -10,11 +10,11 @@ public:
    std::vector<int> row_indices;
    std::vector<int> column_indices;
 
-   COOSymmetricMatrix(int dimension, size_t capacity);
-   COOSymmetricMatrix(int dimension, std::vector<double> matrix, std::vector<int> row_indices, std::vector<int> column_indices);
+   COOSymmetricMatrix(size_t dimension, size_t capacity);
+   COOSymmetricMatrix(size_t dimension, std::vector<double> matrix, std::vector<int> row_indices, std::vector<int> column_indices);
 
    void for_each(const std::function<void (int, int, double)>& f) const override;
-   void insert(double term, int row_index, int column_index) override;
+   void insert(double term, size_t row_index, size_t column_index) override;
    void add_identity_multiple(double multiple) override;
 
    friend std::ostream& operator<<(std::ostream& stream, COOSymmetricMatrix& matrix);
